feat: extract mock data to a seed module
