+++ src/api/routes/courses.js
@@ -1,7 +1,6 @@
 import { Router } from 'express';
 import * as Types from '../../libs/types/common.js';
-import { faker } from '@faker-js/faker';
-import { generateRandomCategory } from './course-categories.js';
+import { generateRandomCourse } from '../../libs/seed.js';
 
 /**
  * @type {Types.Route}
@@ -18,41 +17,3 @@ export default (app) => {
     });
   });
 };
-
-export function generateRandomCourse() {
-  return {
-    id: faker.string.uuid(),
-    name: faker.commerce.product(),
-    code: faker.string.alpha(8),
-    price: faker.commerce.price(),
-    author: faker.person.fullName(),
-    rating: faker.number.float({ min: 0, max: 5, precision: 0.1 }),
-    premium: faker.datatype.boolean(),
-    telegram: faker.internet.url(),
-    onboarding: faker.commerce.productDescription(),
-    difficulty: faker.helpers.arrayElement([
-      'BEGINNER',
-      'INTERMEDIATE',
-      'ADVANCED'
-    ]),
-    description: faker.commerce.productDescription(),
-    intro_video: faker.internet.url(),
-    target_audience: Array.from({ length: 3 }, () =>
-      faker.commerce.productDescription()
-    ),
-    user_id: faker.string.uuid(),
-    course_category_id: faker.string.uuid(),
-    user: {
-      id: faker.string.uuid(),
-      role: faker.helpers.arrayElement(['ADMIN', 'USER']),
-      email: faker.internet.email(),
-      phone_number: faker.helpers.replaceSymbolWithNumber('+628##########'),
-      password: faker.internet.password(),
-      createdAt: faker.date.recent(),
-      updatedAt: faker.date.recent()
-    },
-    course_category: generateRandomCategory(),
-    createdAt: faker.date.recent(),
-    updatedAt: faker.date.recent()
-  };
-}
