+++ src/db/seeders/20231004004355-demo-user.js
@@ -3,7 +3,10 @@ module.exports = {
   up: async (queryInterface, _Sequelize) => {
     const { generateRandomUser } = await import('../../libs/seed.js');
 
-    return queryInterface.bulkInsert('Users', generateRandomUser());
+    return queryInterface.bulkInsert(
+      'Users',
+      Array.from({ length: 100 }, generateRandomUser)
+    );
   },
   down: (queryInterface, _Sequelize) => {
     // @ts-ignore
